requirements/common.txt
ansible==2.1.1.0
appdirs==1.4.0
attrdict==2.0.0
Babel==2.3.4
cffi==1.7.0
cliff==2.2.0
cmd2==0.6.8
colorlog==2.7.0
contextlib2==0.5.4
cryptography==1.5
debtcollector==1.8.0
enum34==1.1.6
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
idna==2.1
ipaddress==1.0.16
iso8601==0.1.11
Jinja2==2.8
jsonpatch==1.14
jsonpointer==1.10
jsonschema==2.5.1
keystoneauth1==2.12.1
MarkupSafe==0.23
monotonic==1.2
msgpack-python==0.4.8
netaddr==0.7.18
netifaces==0.10.5
openstacksdk==0.9.4
os-client-config==1.20.1
osc-lib==1.1.0
oslo.config==3.17.0
oslo.i18n==3.9.0
oslo.serialization==2.13.0
oslo.utils==3.16.0
paramiko==2.0.2
pathlib==1.0.1
pbr==1.10.0
pkg-resources==0.0.0
pollute==1.0.2
positional==1.1.1
prettytable==0.7.2
pyasn1==0.1.9
pycparser==2.14
pycrypto==2.6.1
pyparsing==2.1.8
python-cinderclient==1.8.0
python-glanceclient==2.5.0
python-heatclient==1.4.0
python-keystoneclient==3.5.0
python-novaclient==5.1.0
python-openstackclient==3.1.0
python-swiftclient==3.0.0
pytz==2016.6.1
PyYAML==3.12
requests==2.11.1
requestsexceptions==1.1.3
rfc3986==0.4.1
simplejson==3.8.2
six==1.10.0
stevedore==1.17.1
unicodecsv==0.14.1
warlock==1.2.0
wrapt==1.10.8
